backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# Removed emotion detection dependencies
# opencv-python>=4.8.0
# numpy>=1.24.0
# pillow>=10.0.0
# mediapipe>=0.10.0
# tensorflow>=2.13.0
pydantic>=2.0.0
python-multipart>=0.0.6
openai>=1.0.0
requests>=2.31.0
